requirements.txt
aiohttp==3.8.1
corkus.py==1.2.1
discord.py==1.7.3
matplotlib==3.5.1
psycopg2==2.9.3
PyNaCl==1.5.0
